Move remaining types and function in own modules
Extends ca2c09e
